+++ src/debug.h
@@ -35,7 +35,7 @@ extern void convexFigPrintR(struct convexFigList *list, char *parent);
 extern void convexFigBstPrint(struct convexFigBst *bst);
 extern void convexFigBstPrintR(struct convexFigBst *bst, char *parent, char *attr);
 extern int convexFigBstCheck(struct convexFigBst *bst, char *str);
-extern int convexFigBstCheckR(struct convexFigBst *bst, GLfloat *min, GLfloat *max, int *depth, char *str);
+extern int convexFigBstCheckR(struct convexFigBst *bst, GLdouble *min, GLdouble *max, int *depth, char *str);
 extern void convexLoopDetectPrint(int count, int parent, int child, int index, int next);
 struct trie;
 extern void consoleCmdsTriePrint(struct trie *trie);
